backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
alembic==1.14.0

# Redis
redis==5.2.1

# Embeddings
sentence-transformers==3.3.1
torch>=2.0.0

# LLM APIs
google-generativeai==0.8.3
groq==0.14.0

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Environment and utilities
python-dotenv==1.0.1
httpx==0.28.1
tqdm==4.67.1
unicodedata2==15.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0

# Development
black==24.10.0
isort==5.13.2
mypy==1.14.1
